// src/types/enums.ts
export enum Region {
  NORTH_EAST = "North East",
  NORTH_WEST = "North West",
  NORTH_CENTRAL = "North Central",
  SOUTH_EAST = "South East",
  SOUTH_WEST = "South West",
  SOUTH_SOUTH = "South South",
}

export enum FuelProduct {
  PMS = "PMS",
  AGO = "AGO",
  DPK = "DPK",
  LPG = "LPG",
}

export const FUEL_PRODUCTS = Object.values(FuelProduct);